Optionally use Map type for MP maps.
Accepts Map objects in input and encodes them as MessagePack maps, and
if the decoder option "usemap" is true, it decodes maps to Map objects
instead of plain JS Objects.

The key advantage of this is that Maps, like MessagePack maps, and
unlike JS Objects, can have keys of arbitrary type.